// web/src/enums/events.ts
export enum Send {
  saveOutfit = 'saveOutfit',
  renameOutfit = 'renameOutfit',
  deleteOutfit = 'deleteOutfit',
  useOutfit = 'useOutfit',
  importOutfit = 'importOutfit',
  itemOutfit = 'itemOutfit',
  setTattoos = 'setTattoos',
  cancel = 'cancel',
  save = 'save',
  setModel = 'setModel',
  getModelTattoos = 'getModelTattoos',
  setHeadBlend = 'setHeadBlend',
  setHeadStructure = 'setHeadStructure',
  setHeadOverlay = 'setHeadOverlay',
  setProp = 'setProp',
  setDrawable = 'setDrawable',
  toggleItem = 'toggleItem',
}

export enum Receive {
  data = 'data',
}